requirements.txt
django==3.0.11
djangorestframework==3.12.2
django-filter==2.3.0
django-cors-headers==3.5.0
django-environ==0.4.5
pytils==0.3
Pillow==8.0.1
django-imagekit==4.0.2
django-taggit==1.3.0
requests==2.25.0
wikipedia==1.4.0
django-debug-toolbar==3.1.1